Make balloons pop when they touch spikey things.
The matter physics engine doesn't have great support for collision
handling, youy get a single `onCollideStart` callback and have to
figure out if it was called on the objects you're interested in.

Instead I've added `PhaserMatterCollisionPlugin` which allows us to
specifically create collision callbacks on pairs of objects.

Also started clening up some of the object creation in the game scene
and removing old assets we're not using.