Make the balloon string silky smooth
This is a little hackier than I'd like, but the end result is actually
pretty goood.

We store the rope sections as an array on the scene and in the
`update` method we construct an array of `Phaser.Math.Vector2` objects
based on the coordinates of the rope sections.

We can then use the `Phaser.Curves` namespace to draw a curve through
these vectors.

I've set the rope sections (except the rope anchor) to be invisible :)